add some change

+++ Learning/logic.py
@@ -103,21 +103,22 @@ def get_analysis_context(page, item):
 def get_delete_context(page, item):
     if models.ModelInfo.objects.filter(id=item):
         obj = models.ModelInfo.objects.get(id=item)
-        obj.delete()
-        extension = {
-            'model': '.pth',
-            'log': '.log',
-            'loss_curve': '.png',
-            'accuracy_curve': '.png',
-        }
-        for sub_save_path in constants.sub_save_paths:
-            try:
-                os.remove(
-                    constants.extra_args['save_path'] + os.sep + constants.extra_args[sub_save_path] + os.sep + str(
-                        item) +
-                    extension[sub_save_path])
-            except FileNotFoundError:
-                pass
+        if not obj.state == 'train':
+            obj.delete()
+            extension = {
+                'model': '.pth',
+                'log': '.log',
+                'loss_curve': '.png',
+                'accuracy_curve': '.png',
+            }
+            for sub_save_path in constants.sub_save_paths:
+                try:
+                    os.remove(
+                        constants.extra_args['save_path'] + os.sep + constants.extra_args[sub_save_path] + os.sep + str(
+                            item) +
+                        extension[sub_save_path])
+                except FileNotFoundError:
+                    pass
     context = {
         'redirect': '/result/' + str(page) + '/',
     }
@@ -133,7 +134,7 @@ def get_clean_context():
                 item = int(reg.match(filename).group(1))
                 if not models.ModelInfo.objects.filter(id=item):
                     os.remove(constants.extra_args['save_path'] + os.sep + sub_save_path + os.sep + filename)
-            except Exception as e:
+            except ValueError or FileNotFoundError:
                 pass
     context = {
         'redirect': '/result/1/',
